reintroduce n g_is
The security of using only a single g value is unstudied for the case of
attacking a system with no low-level encodings of zero (per communication with
Jung Hee), and so to be extra careful we revert back to using a vector if g
values.  The only real hit here is slightly larger secret key and slightly
slower generation times.  But because the g primes are O(\lambda), the hit isn't
that much.