Don't choose n = λ
I was finally convinced that setting n = λ is A BAD IDEA, so we need to set it
to a larger value now.  I chose to use the estimates in Tancrède Lepoint's
thesis (§7.2.4), looping through choosing n, η, and ν until a fixed point is
reached (as they all depend on each other).  Note that the estimates on LLL and
BKZ difficulty from Tancrède's thesis are around 5 years old, so likely we are
UNDERESTIMATING the security.  This should be fixed!